Use PIMPL so that the interpolant can be passed as a functor.
